+++ iptables/metadata.rb
@@ -1,7 +1,7 @@
-name             'ansible'
+name             'iptables'
 maintainer       'JMLRT'
 maintainer_email 'julien@mailleret.fr'
 license          'All rights reserved'
-description      'Installs/Configures ansible'
+description      'Installs/Configures iptables'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'