+++ vim/metadata.rb
@@ -1,6 +1,6 @@
 name             'vim'
 maintainer       'JMLRT'
-maintainer_email 'YOUR_EMAIL'
+maintainer_email 'julien@mailleret.fr'
 license          'All rights reserved'
 description      'Installs/Configures vim'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))